Add function/type utilities for parsing static URLs
Prepare a function to handle the conversion of a URL property name
(provided by the `.yml` content files) to the static URLs configured
in `nuxt.config.ts`. Also add a union type for the `/goods` content, as
this adds a `version: string` property that does not already exist on
the nuxt `PageLink` type.