Update Explorer URL and add static URL config (#5)
* Add static URLs to `nuxt.config.ts`

Placeholder object for all static Lotusia web resources. For use later
via `useRuntimeConfig()`.

https://nuxt.com/docs/guide/going-further/runtime-config

* Add function/type utilities for parsing static URLs

Prepare a function to handle the conversion of a URL property name
(provided by the `.yml` content files) to the static URLs configured
in `nuxt.config.ts`. Also add a union type for the `/goods` content, as
this adds a `version: string` property that does not already exist on
the nuxt `PageLink` type.

* Apply static links from `nuxt.config.ts`

Inserts the `nuxt.config.ts` URL properties into the `.yml` files, and
then imports and runs the `parsePageLinks()` util function to create the
valid links on the pages.