Bugfix: Missing \`post\` object or \`postId\` in API results (#25)
* Bugfix: Profile vote history table doesn't render

RANK votes are not required to have a `postId` defined, and this was causing the "Vote History" table to not render the API results. This commit adds a conditional to the Vue template to show votes as profile RANK votes when there is no `post` object available.

* Bugfix: No `postId` on "Latest Activity" page

Fixes the same bug as 76bca46, but on the "Latest Activity" page