A few updates (#33)
This branch fixes a HTTP 500 with the new Explorer related to the GeoIP
module. Also, with the recent change from `lotus-lib` to `lotus-sdk`,
the MuSig2 blog article and Taproot documentation was updated.

A recent address encoding issue was also discovered in `lotus-sdk` that
changes the XAddress format, so the documentation was also updated to
coincide with this fix.